+++ pkger/clone_resource.go
@@ -641,6 +641,18 @@ func convertCellView(cell influxdb.Cell) chart {
 		setNoteFixes(p.Note, p.ShowNoteWhenEmpty, p.Prefix, p.Suffix)
 		ch.TickPrefix = p.TickPrefix
 		ch.TickSuffix = p.TickSuffix
+	case influxdb.MosaicViewProperties:
+		ch.Kind = chartKindMosaic
+		ch.Queries = convertQueries(p.Queries)
+		ch.Colors = stringsToColors(p.ViewColors)
+		ch.XCol = p.XColumn
+		ch.YSeriesColumns = p.YSeriesColumns
+		ch.Axes = []axis{
+			{Label: p.XAxisLabel, Prefix: p.XPrefix, Suffix: p.XSuffix, Name: "x", Domain: p.XDomain},
+			{Label: p.YAxisLabel, Prefix: p.YPrefix, Suffix: p.YSuffix, Name: "y", Domain: p.YDomain},
+		}
+		ch.Note = p.Note
+		ch.NoteOnEmpty = p.ShowNoteWhenEmpty
 	case influxdb.ScatterViewProperties:
 		ch.Kind = chartKindScatter
 		ch.Queries = convertQueries(p.Queries)
@@ -696,14 +708,23 @@ func convertChartToResource(ch chart) Resource {
 		fieldChartHeight: ch.Height,
 		fieldChartWidth:  ch.Width,
 	}
-	if len(ch.Queries) > 0 {
-		r[fieldChartQueries] = ch.Queries
+	var qq []Resource
+	for _, q := range ch.Queries {
+		qq = append(qq, Resource{
+			fieldQuery: q.DashboardQuery(),
+		})
+	}
+	if len(qq) > 0 {
+		r[fieldChartQueries] = qq
 	}
 	if len(ch.Colors) > 0 {
 		r[fieldChartColors] = ch.Colors
 	}
 	if len(ch.Axes) > 0 {
 		r[fieldChartAxes] = ch.Axes
+	}
+	if len(ch.YSeriesColumns) > 0 {
+		r[fieldChartYSeriesColumns] = ch.YSeriesColumns
 	}
 	if ch.EnforceDecimals {
 		r[fieldChartDecimalPlaces] = ch.DecimalPlaces